restored original (Oleg) version of ssax/SSAX-code.rkt   w/ disabled run-test and SSAX:XML->SXML (all-caps name)
The code this deletes and replaces was the same as the original version
except that the defn and all uses of run-test were removed,
and the defn of SSAX:XML->SXML (backwards-compat allcaps name) was removed.
(I checked, using a combination of pretty-write and diff)

The benefit: restored comments, restored test cases.
Next step: extract tests.